+++ examples/keyspace.rs
@@ -8,8 +8,7 @@ use tokio::time::sleep;
 async fn fake_traffic(client: &RedisClient, amount: usize) -> Result<(), RedisError> {
   // use a new client since the provided client is subscribed to keyspace events
   let client = client.clone_new();
-  let _ = client.connect();
-  client.wait_for_connect().await?;
+  client.init().await?;
 
   for idx in 0 .. amount {
     let key: RedisKey = format!("foo-{}", idx).into();
@@ -58,8 +57,7 @@ async fn centralized_keyspace_events() -> Result<(), RedisError> {
   });
 
   // connect after setting up the reconnection logic
-  subscriber.connect();
-  subscriber.wait_for_connect().await?;
+  subscriber.init().await?;
 
   let mut keyspace_rx = subscriber.keyspace_event_rx();
   // set up a task that listens for keyspace events
@@ -107,8 +105,7 @@ async fn clustered_keyspace_events() -> Result<(), RedisError> {
   });
 
   // connect after setting up the reconnection logic
-  subscriber.connect();
-  subscriber.wait_for_connect().await?;
+  subscriber.init().await?;
 
   let mut keyspace_rx = subscriber.keyspace_event_rx();
   // set up a task that listens for keyspace events